fix: исправления базовой реализации ТЗ №9.
