+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -1,15 +1,17 @@
 package ru.yandex.practicum.filmorate.model;
 
-import lombok.Data;
+import lombok.*;
 
 import java.time.Duration;
 import java.time.LocalDate;
 
 @Data
+@RequiredArgsConstructor
+@AllArgsConstructor
 public class Film {
     private final int id;
     private String name;
     private String description;
     private LocalDate releaseDate;
-    private Duration duration;
+    private int duration;
 }
